+++ projects/net.splitcells.dem.api/CHANGELOG.md
@@ -89,6 +89,7 @@
      This also makes it possible to use the name `Resource` for an interface
      of resource like things.
 ### Minor Changes
+* **2025-11-16\#42** `Dem#process` now supports `Cell#programName`, as an alternative ProgramName.
 * **2025-09-22\#37** Define helper methods `hasKey`, `hasNotkey` and `value` for Map, in order to define more type safe access.
 * **2025-09-22\#37** Define Error Reporting API.
 * **2025-09-04\#37** Define Needs Check API.